back_end/requirements.txt
autopep8==1.5.5
blinker==1.4
cachelib==0.1.1
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.10
Flask-Mail==0.9.1
Flask-Session==0.3.2
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
Pillow==8.1.1
pycodestyle==2.6.0
reportlab==3.5.62
six==1.15.0
toml==0.10.2
Werkzeug==1.0.1
gunicorn==20.0.4
